+++ chainercv/links/model/yolo/__init__.py
@@ -1,2 +1,3 @@
-from chainercv.links.model.yolo.yolov3 import ResidualBlock  # NOQA
-from chainercv.links.model.yolo.yolov3 import YOLOv3  # NOQA
+from chainercv.links.model.yolo.yolo_v3 import Darknet53Extractor  # NOQA
+from chainercv.links.model.yolo.yolo_v3 import ResidualBlock  # NOQA
+from chainercv.links.model.yolo.yolo_v3 import YOLOv3  # NOQA
